ref: free_pipe_attr関数を専用ファイルに移動
