+++ command/src/pipe_process.c
@@ -6,7 +6,7 @@
 /*   By: tkirihar <tkirihar@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/10 21:07:42 by tkirihar          #+#    #+#             */
-/*   Updated: 2022/02/23 22:44:12 by tkirihar         ###   ########.fr       */
+/*   Updated: 2022/02/23 22:48:44 by tkirihar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -174,26 +174,6 @@ void	exec_cmd(t_exec_attr *ea, t_pipe_attr *pa)
 	{
 		close_pipe(pa);
 	}
-}
-
-void	free_pipe_fd(t_pipe_attr *pa)
-{
-	int	i;
-
-	i = 0;
-	while (i < pa->pipe_count)
-	{
-		free(pa->pipe_fd[i]);
-		i++;
-	}
-	free(pa->pipe_fd);
-}
-
-void	free_pipe_attr(t_pipe_attr *pa)
-{
-	// ここでpa->current_cmdもfreeすべきかも
-	free_pipe_fd(pa);
-	free(pa->cpid_array);
 }
 
 void	pipe_process(t_exec_attr *ea, int pipe_count)
